- Removed aconf->ip = IN_ADDR(ip); warnings (force unsigned long). - Commented out valid_comment function (unused so far?).

+++ modules/m_kline.c
@@ -20,7 +20,7 @@
  *   along with this program; if not, write to the Free Software
  *   Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  *
- *   $Id: m_kline.c,v 1.54 2001/01/24 20:04:38 fl_ Exp $
+ *   $Id: m_kline.c,v 1.55 2001/01/31 04:53:21 jdc Exp $
  */
 #include "tools.h"
 #include "m_kline.h"
@@ -88,7 +88,13 @@ static char *cluster(char *);
 static int find_user_host(struct Client *sptr,
                           char *user_host_or_nick, char *user, char *host);
 
+/*
+ * jdc -- unused?
+ */
+/*
 static int valid_comment(struct Client *sptr, char *comment);
+*/
+
 static int valid_user_host(struct Client *sptr, char *user, char *host);
 static int valid_wild_card(struct Client *sptr, char *user, char *host);
 static int already_placed_kline(struct Client *sptr, char *user, char *host,
@@ -228,7 +234,7 @@ static int mo_kline(struct Client *cptr,
 
   if (ip_kline)
    { 
-     aconf->ip = IN_ADDR(ip);
+     aconf->ip = (unsigned long) IN_ADDR(ip);
      aconf->ip_mask = ip_mask;
    }
 
@@ -332,7 +338,7 @@ static int ms_kline(struct Client *cptr,
 
       if(ip_kline)
         {
-          aconf->ip = IN_ADDR(ip);
+          aconf->ip = (unsigned long) IN_ADDR(ip);
           aconf->ip_mask = ip_mask;
         }
 
@@ -361,7 +367,7 @@ static void apply_kline(struct Client *sptr, struct ConfItem *aconf,
 {
   if(ip_kline)
     {
-      aconf->ip = PIN_ADDR(ip);
+      aconf->ip = (unsigned long) PIN_ADDR(ip);
       aconf->ip_mask = ip_mask;
       add_ip_Kline(aconf);
     }
@@ -943,6 +949,10 @@ static int valid_wild_card(struct Client *sptr, char *luser, char *lhost)
  * output       - 0 if no valid comment, 1 if valid
  * side effects - NONE
  */
+/*
+ * jdc -- unused?
+ */
+/*
 static int valid_comment(struct Client *sptr, char *comment)
 {
   if(strchr(comment, ':'))
@@ -964,6 +974,7 @@ static int valid_comment(struct Client *sptr, char *comment)
     }
   return 1;
 }
+*/
 
 /*
  * already_placed_kline
